+++ src/databricks/labs/ucx/source_code/dependency_graph.py
@@ -3,50 +3,16 @@ from __future__ import annotations
 import abc
 import ast
 import typing
-from dataclasses import dataclass
 from collections.abc import Callable
 from pathlib import Path
 
-
+from databricks.labs.ucx.source_code.dependency_problem import DependencyProblem
 from databricks.labs.ucx.source_code.python_linter import ASTLinter, PythonLinter
 
 if typing.TYPE_CHECKING:
     from databricks.labs.ucx.source_code.dependency_resolvers import DependencyResolver
-    from databricks.labs.ucx.source_code.dependency_loaders import SourceContainer, DependencyLoader
-
-
-MISSING_SOURCE_PATH = "<MISSING_SOURCE_PATH>"
-
-
-@dataclass
-class DependencyProblem:
-    code: str
-    message: str
-    source_path: Path = Path(MISSING_SOURCE_PATH)
-    start_line: int = -1
-    start_col: int = -1
-    end_line: int = -1
-    end_col: int = -1
-
-    def replace(
-        self,
-        code: str | None = None,
-        message: str | None = None,
-        source_path: Path | None = None,
-        start_line: int | None = None,
-        start_col: int | None = None,
-        end_line: int | None = None,
-        end_col: int | None = None,
-    ) -> DependencyProblem:
-        return DependencyProblem(
-            code if code is not None else self.code,
-            message if message is not None else self.message,
-            source_path if source_path is not None else self.source_path,
-            start_line if start_line is not None else self.start_line,
-            start_col if start_col is not None else self.start_col,
-            end_line if end_line is not None else self.end_line,
-            end_col if end_col is not None else self.end_col,
-        )
+    from databricks.labs.ucx.source_code.dependency_containers import SourceContainer
+    from databricks.labs.ucx.source_code.dependency_loaders import DependencyLoader
 
 
 class Dependency(abc.ABC):
