+++ src/databricks/labs/ucx/source_code/notebook_migrator.py
@@ -5,7 +5,7 @@ from databricks.sdk.service.workspace import ExportFormat, ObjectInfo, ObjectTyp
 
 from databricks.labs.ucx.source_code.languages import Languages
 from databricks.labs.ucx.source_code.notebook import Notebook, RunCell
-from databricks.labs.ucx.source_code.dependencies import (
+from databricks.labs.ucx.source_code.dependency_graph import (
     Dependency,
 )
 from databricks.labs.ucx.source_code.dependency_loaders import WorkspaceNotebookLoader
