+++ tests/unit/__init__.py
@@ -17,7 +17,8 @@ from databricks.sdk.service.sql import EndpointConfPair
 from databricks.sdk.service.workspace import ExportResponse, GetSecretResponse, Language
 
 from databricks.labs.ucx.hive_metastore.mapping import TableMapping, TableToMigrate
-from databricks.labs.ucx.source_code.dependency_loaders import SourceContainer, LocalFileLoader
+from databricks.labs.ucx.source_code.dependency_containers import SourceContainer
+from databricks.labs.ucx.source_code.dependency_loaders import LocalFileLoader
 from databricks.labs.ucx.source_code.files import LocalFile
 from databricks.labs.ucx.source_code.notebook import Notebook
 from databricks.labs.ucx.source_code.base import NOTEBOOK_HEADER
