+++ tests/unit/source_code/test_s3fs.py
@@ -2,11 +2,12 @@ from pathlib import Path
 
 import pytest
 
-from databricks.labs.ucx.source_code.dependencies import (
+from databricks.labs.ucx.source_code.dependency_graph import (
     DependencyGraphBuilder,
-    DependencyProblem,
 )
-from databricks.labs.ucx.source_code.dependency_loaders import SourceContainer, LocalNotebookLoader, LocalFileLoader
+from databricks.labs.ucx.source_code.dependency_problem import DependencyProblem
+from databricks.labs.ucx.source_code.dependency_loaders import LocalFileLoader, LocalNotebookLoader
+from databricks.labs.ucx.source_code.dependency_containers import SourceContainer
 from databricks.labs.ucx.source_code.dependency_resolvers import DependencyResolver
 from databricks.labs.ucx.source_code.site_packages import SitePackages
 from databricks.labs.ucx.source_code.syspath_provider import SysPathProvider
